Updated to 7.1 moved blocks and items to their own classes and fixed all errors that came with it
