+++ gradle.properties
@@ -7,7 +7,9 @@ bta_version=7.1-pre1a
 loader_version=0.14.19-babric.3-bta
 
 # HalpLibe
+terrain_api_version=1.4.2-7.1
 halplibe_version=3.1.0
+dragonfly_version=1.3.1-7.1
 
 # Mod
 mod_version=1.0.0-7.1
