+++ src/main/java/bta/aether/entity/EntityArrowFlaming.java
@@ -1,5 +1,6 @@
 package bta.aether.entity;
 
+import bta.aether.item.AetherItems;
 import com.mojang.nbt.CompoundTag;
 import net.minecraft.client.render.Tessellator;
 import net.minecraft.core.HitResult;
@@ -36,6 +37,7 @@ public class EntityArrowFlaming extends Entity {
     protected float arrowGravity;
     protected int arrowDamage;
     protected int arrowType;
+    protected ItemStack stack;
 
     public EntityArrowFlaming(World world) {
         this(world, 4);
@@ -43,6 +45,7 @@ public class EntityArrowFlaming extends Entity {
 
     public EntityArrowFlaming(World world, int arrowType) {
         super(world);
+        this.stack = new ItemStack(Item.ammoArrow);
         this.xTile = -1;
         this.yTile = -1;
         this.zTile = -1;
@@ -58,6 +61,7 @@ public class EntityArrowFlaming extends Entity {
 
     public EntityArrowFlaming(World world, double d, double d1, double d2, int arrowType) {
         super(world);
+        this.stack = new ItemStack(Item.ammoArrow);
         this.xTile = -1;
         this.yTile = -1;
         this.zTile = -1;
@@ -75,6 +79,7 @@ public class EntityArrowFlaming extends Entity {
 
     public EntityArrowFlaming(World world, EntityLiving entityliving, boolean doesArrowBelongToPlayer, int arrowType) {
         super(world);
+        this.stack = new ItemStack(Item.ammoArrow);
         this.xTile = -1;
         this.yTile = -1;
         this.zTile = -1;
@@ -343,10 +348,13 @@ public class EntityArrowFlaming extends Entity {
 
     public void playerTouch(EntityPlayer player) {
         if (!this.world.isClientSide) {
-            if (this.inGround && this.doesArrowBelongToPlayer && this.arrowShake <= 0 && player.inventory.addItemStackToInventory(new ItemStack(Item.ammoArrow, 1))) {
-                this.world.playSoundAtEntity(this, "random.pop", 0.2F, ((this.random.nextFloat() - this.random.nextFloat()) * 0.7F + 1.0F) * 2.0F);
-                player.onItemPickup(this, Item.ammoArrow.id);
-                this.remove();
+            if (this.inGround && this.doesArrowBelongToPlayer && this.arrowShake <= 0) {
+                player.inventory.insertItem(this.stack, true);
+                if (this.stack.stackSize <= 0) {
+                    this.world.playSoundAtEntity(this, "random.pop", 0.2F, ((this.random.nextFloat() - this.random.nextFloat()) * 0.7F + 1.0F) * 2.0F);
+                    player.onItemPickup(this, Item.ammoArrow.id);
+                    this.remove();
+                }
             }
 
         }
