+++ src/main/java/bta/aether/world/BiomeAether.java
@@ -1,6 +1,7 @@
 package bta.aether.world;
 
 import bta.aether.Aether;
+import bta.aether.block.AetherBlocks;
 import net.minecraft.core.world.biome.Biome;
 import net.minecraft.core.world.generate.feature.WorldFeature;
 import net.minecraft.core.world.generate.feature.tree.WorldFeatureTree;
@@ -17,9 +18,9 @@ public class BiomeAether
 
         if(random.nextInt(10) == 0)
         {
-            return new WorldFeatureTreeFancy(Aether.leavesSkyroot.id, Aether.logSkyroot.id);
+            return new WorldFeatureTreeFancy(AetherBlocks.leavesSkyroot.id, AetherBlocks.logSkyroot.id);
         } else
-            return new WorldFeatureTree(Aether.leavesSkyroot.id, Aether.logSkyroot.id, 5);
+            return new WorldFeatureTree(AetherBlocks.leavesSkyroot.id, AetherBlocks.logSkyroot.id, 5);
 
     }
 }
