+++ src/main/resources/lang/aether/en_US.lang
@@ -57,9 +57,9 @@ tile.aether.block.zanite.desc=A solid block of zanite. Shiny!
 tile.aether.trap.name=Trap
 tile.aether.trap.desc=Sneaky block that summons a sentry
 tile.aether.chest.mimic.name=Trapped Chest
-tile.aether.chest.mimic.desc=Thank you dark souls...
+tile.aether.chest.mimic.desc=Thank you dark souls.
 tile.aether.chest.treasure.name=Treasure Chest
-tile.aether.chest.treasure.desc=A chest filled with mysterious loot...
+tile.aether.chest.treasure.desc=A chest filled with mysterious loot.
 tile.aether.carved.name=Carved Stone
 tile.aether.carved.desc=Decorative bronze dungeon block.
 tile.aether.angelic.name=Angelic Stone
